+++ controllers/tasks.py
@@ -2,33 +2,18 @@ import os
 import json
 import uuid
 
-from google.cloud import tasks_v2, bigquery  # type: ignore
+from google.cloud import tasks_v2
 from google import auth
 
+from libs.bigquery import get_customers
+
+TASKS_CLIENT = tasks_v2.CloudTasksClient()
 _, PROJECT_ID = auth.default()
 CLOUD_TASKS_PATH = (PROJECT_ID, "us-central1", "luux_media_emails")
 
 
-def get_customers(
-    client: bigquery.Client,
-    dataset: str,
-    table_suffix: str,
-) -> list[str]:
-    results = client.query(
-        f"SELECT DISTINCT ExternalCustomerId FROM {dataset}.AccountStats_{table_suffix}"
-    ).result()
-    rows = [dict(row.items()) for row in results]
-    return [str(i["ExternalCustomerId"]) for i in rows]
-
-
-def create_tasks(
-    task_client: tasks_v2.CloudTasksClient,
-    bq_client: bigquery.Client,
-    dataset: str,
-    table_suffix: str,
-    tasks_data: dict,
-) -> dict:
-    accounts = get_customers(bq_client, dataset, table_suffix)
+def create_tasks(dataset: str, table_suffix: str, tasks_data: dict) -> dict:
+    accounts = get_customers(dataset, table_suffix)
     payloads = [
         {
             "name": f"{account}-{uuid.uuid4()}",
@@ -41,12 +26,14 @@ def create_tasks(
     ]
     tasks = [
         {
-            "name": task_client.task_path(*CLOUD_TASKS_PATH, task=str(payload["name"])),
+            "name": TASKS_CLIENT.task_path(
+                *CLOUD_TASKS_PATH, task=str(payload["name"])
+            ),
             "http_request": {
                 "http_method": tasks_v2.HttpMethod.POST,
                 "url": os.getenv("PUBLIC_URL"),
                 "oidc_token": {
-                    "service_account_email": os.getenv('GCP_SA'),
+                    "service_account_email": os.getenv("GCP_SA"),
                 },
                 "headers": {"Content-type": "application/json"},
                 "body": json.dumps(payload["payload"]).encode(),
@@ -57,9 +44,9 @@ def create_tasks(
     return {
         "messages_sent": len(
             [
-                task_client.create_task(
+                TASKS_CLIENT.create_task(
                     request={
-                        "parent": task_client.queue_path(*CLOUD_TASKS_PATH),
+                        "parent": TASKS_CLIENT.queue_path(*CLOUD_TASKS_PATH),
                         "task": task,
                     }
                 )
